refactor: move ModuleMatcher to xonsh.lib
